Develop

+++ app/Http/Controllers/User/ClassController.php
@@ -9,22 +9,30 @@ use App\Models\User;
 use Illuminate\Support\Facades\Log;
 use Yajra\DataTables\Facades\DataTables;
 use App\Http\Requests\Class\StoreClassRequest;
+use App\Models\Major;
+use App\Models\Teacher;
 
 class ClassController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware(['role:vice-principal']);
-    }
-
     public function index(Request $request)
     {
+        if (!$request->user()->can('class.*')) {
+            return abort(403);
+        }
+
         if ($request->ajax()) {
-            $data = SchoolClass::filter($request->all());
+            $data = SchoolClass::query()
+                ->leftJoin('majors', 'classes.major_id', '=', 'majors.id')
+                ->select([
+                    'classes.id',
+                    'classes.name',
+                    'classes.level',
+                    'majors.name as major_name',
+                ])
+                ->filter($request->all());
 
             return Datatables::of($data)
                 ->addColumn('id', function ($row) {
-                    // Memperbaiki kode: id checkbox harus unik, gunakan id dinamis dan name array agar bisa multiple select
                     $html = '
                         <div class="checkbox-checked">
                             <div class="form-check d-flex justify-content-center align-items-center">
@@ -34,9 +42,6 @@ class ClassController extends Controller
                         </div>
                         ';
                     return $html;
-                })
-                ->editColumn('created_at', function ($row) {
-                    return $row->created_at;
                 })
                 ->addColumn('Nama', function ($row) {
                     $html = '
@@ -54,48 +59,32 @@ class ClassController extends Controller
                 })
                 ->addColumn('Jurusan', function ($row) {
                     $html = '
-                        <span class="badge badge-light-primary">' . $row->major . '</span>
-                    ';
-                    return $html;
-                })
-                ->addColumn('Kapasitas', function ($row) {
-                    $html = '
-                        <p class="f-light">' . $row->capacity . '</p>
+                        <span class="badge badge-light-primary">' . ($row->major_name ? $row->major_name : " - ") . '</span>
                     ';
                     return $html;
                 })
                 ->addColumn('Aksi', function ($row) {
                     $html = '
                     <div class="common-align gap-2 justify-content-start">
-                        <a class="square-white edit" href="#!" data-id="' . $row->id . '">
+                        <a class="square-white edit" style="cursor: pointer" data-id="' . $row->id . '">
                             <svg><use href="' . asset('assets/svg/icon-sprite.svg#edit-content') . '"></use></svg>
                         </a>
-                        <a class="square-white trash" href="#!" data-id="' . $row->id . '">
+                        <a class="square-white trash" style="cursor: pointer" data-id="' . $row->id . '">
                             <svg><use href="' . asset('assets/svg/icon-sprite.svg#trash1') . '"></use></svg>
                         </a>
                     </div>';
                     return $html;
                 })
-                ->rawColumns(['id', 'Nama', 'Tingkat', 'Jurusan',  'Kapasitas', 'Aksi'])
+                ->rawColumns(['id', 'Nama', 'Tingkat', 'Jurusan', 'Aksi'])
                 ->make(true);
         } else {
-            $classes = SchoolClass::filter(request()->all())->paginate(10);
             $levels = SchoolClass::select('level')->distinct()->get();
-            $majors = SchoolClass::select('major')->distinct()->get();
+            $majors = Major::select('id', 'name')->get();
             return view('user.class.index', [
-                'classes' => $classes,
                 'levels' => $levels,
-                'majors' => $majors
+                'majors' => $majors,
             ]);
         }
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
     }
 
     /**
@@ -104,24 +93,22 @@ class ClassController extends Controller
     public function store(StoreClassRequest $request)
     {
         try {
-            // Buat kelas baru
-            $schoolClass = SchoolClass::create([
-                'name' => $request->validated('name'),
-                'level' => $request->validated('level'),
-                'major' => $request->validated('major'),
-                'capacity' => $request->validated('capacity')
-            ]);
+            if (!$request->user()->can('class.*')) {
+                return abort(403);
+            }
 
-            // Response sukses
+            $validated = $request->validated();
+
+            $class = SchoolClass::create($validated);
+
             return $this->sendResponse(
                 'Kelas berhasil ditambahkan',
                 [
-                    'id' => $schoolClass->id,
-                    'name' => $schoolClass->name,
-                    'level' => $schoolClass->level,
-                    'major' => $schoolClass->major,
-                    'capacity' => $schoolClass->capacity,
-                    'created_at' => $schoolClass->created_at->translatedFormat('d/m/Y H:i')
+                    'id' => $class->id,
+                    'name' => $class->name,
+                    'level' => $class->level,
+                    'major' => $class->major->name ?? '-',
+                    'created_at' => $class->created_at->translatedFormat('d/m/Y H:i')
                 ],
                 201
             );
@@ -137,9 +124,13 @@ class ClassController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show($id)
+    public function edit(Request $request, $id)
     {
         try {
+            if (!$request->user()->can('class.*')) {
+                return abort(403);
+            }
+
             $class = SchoolClass::find($id);
 
             if (!$class) {
@@ -161,24 +152,24 @@ class ClassController extends Controller
     }
 
     /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
-    }
-
-    /**
      * Update the specified resource in storage.
      */
     public function update(StoreClassRequest $request, $id)
     {
         try {
+            if (!$request->user()->can('class.*')) {
+                return abort(403);
+            }
+
             $class = SchoolClass::findOrFail($id);
-            $class->update($request->validated());
+
+            $validated = $request->validated();
+
+            $class->update($validated);
 
             return $this->sendResponse('Kelas berhasil diedit', $class);
         } catch (\Exception $e) {
+            Log::error($e);
             return $this->sendError(
                 'Silakan coba lagi.',
                 [],
@@ -193,7 +184,25 @@ class ClassController extends Controller
     public function destroy(Request $request, $id)
     {
         try {
+            if (!$request->user()->can('class.*')) {
+                return abort(403);
+            }
+
             $class = SchoolClass::findOrFail($id);
+            if ($class->students->count() > 0) {
+                return $this->sendError(
+                    'Kelas tidak dapat dihapus karena masih memiliki siswa.',
+                    [],
+                    400
+                );
+            }
+            if ($class->schedules->count() > 0) {
+                return $this->sendError(
+                    'Kelas tidak dapat dihapus karena masih memiliki jadwal.',
+                    [],
+                    400
+                );
+            }
             $class->delete();
 
             return $this->sendResponse(
@@ -211,6 +220,9 @@ class ClassController extends Controller
     public function bulkDestroy(Request $request)
     {
         try {
+            if (!$request->user()->can('class.*')) {
+                return abort(403);
+            }
 
             $ids = $request->input('ids');
 
@@ -222,7 +234,24 @@ class ClassController extends Controller
                 );
             }
 
-            SchoolClass::whereIn('id', $ids)->delete();
+            $classes = SchoolClass::whereIn('id', $ids)->get();
+            foreach ($classes as $class) {
+                if ($class->students->count() > 0) {
+                    return $this->sendError(
+                        'Kelas tidak dapat dihapus karena masih memiliki siswa.',
+                        [],
+                        400
+                    );
+                }
+                if ($class->schedules->count() > 0) {
+                    return $this->sendError(
+                        'Kelas tidak dapat dihapus karena masih memiliki jadwal.',
+                        [],
+                        400
+                    );
+                }
+                $class->delete();
+            }
 
             return $this->sendResponse(
                 'Data yang dipilih berhasil dihapus.'
